Move HeterogeneousProduct to its own package
