--- HeterogeneousCore/AcceleratorService/src/classes.h
@@ -1,9 +1,0 @@
-#include "DataFormats/Common/interface/Wrapper.h"
-#include "HeterogeneousCore/AcceleratorService/interface/HeterogeneousProduct.h"
-
-namespace {
-    struct dictionary {
-      HeterogeneousProduct<unsigned int, unsigned int> hpuu;
-      HeterogeneousProduct<unsigned int, int*> hpuup;
-    };
-}
