+++ HeterogeneousCore/AcceleratorService/test/BuildFile.xml
@@ -1,13 +1,10 @@
-<bin name="testHeteroGeneousProduct" file="testHeterogeneousProduct.cpp">
-  <use name="cppunit"/>
-  <use name="HeterogeneousCore/AcceleratorService"/>
-</bin>
 <library file="TestAcceleratorServiceProducerGPUMock.cc" name="HeterogeneousCoreAcceleratorServiceTest">
   <flags EDM_PLUGIN="1"/>
   <use name="FWCore/Framework"/>
   <use name="FWCore/PluginManager"/>
   <use name="FWCore/ParameterSet"/>
   <use name="HeterogeneousCore/AcceleratorService"/>
+  <use name="HeterogeneousCore/Product"/>
 </library>
 <library file="TestAcceleratorServiceProducerGPU.cc TestAcceleratorServiceProducerGPUHelpers.cu" name="HeterogeneousCoreAcceleratorServiceTestGPU">
   <flags EDM_PLUGIN="1"/>
@@ -15,6 +12,7 @@
   <use name="FWCore/PluginManager"/>
   <use name="FWCore/ParameterSet"/>
   <use name="HeterogeneousCore/AcceleratorService"/>
+  <use name="HeterogeneousCore/Product"/>
   <use name="cuda"/>
   <use name="cuda-api-wrappers"/>
 </library>
