+++ HeterogeneousCore/Product/BuildFile.xml
@@ -1,0 +1,7 @@
+<use name="DataFormats/Common"/>
+<use name="FWCore/Utilities"/>
+<use name="root"/>
+
+<export>
+    <lib name="1"/>
+</export>
