+++ HeterogeneousCore/Product/src/classes_def.xml
